refactor: move video folder to media
