+++ src/components/index.ts
@@ -7,6 +7,8 @@ import OpenLayersMap from './map/openLayersMap'
 import MarkedMarkdown from './markdown/markedMarkdown'
 import MarkedStreamingMarkdown from './markdown/markedStreamingMarkdown'
 import Sound from './media/audio/sound'
+import Video from './media/video/video'
+import YoutubeVideo from './media/video/youtubeVideo'
 import PopoverMenu from './menu/popoverMenu'
 import MessageCanvas from './messageCanvas/messageCanvas'
 import MessageSpace from './messageSpace/messageSpace'
@@ -17,8 +19,6 @@ import StreamingText from './text/streamingText'
 import Text from './text/text'
 import TextInput from './textInput/textInput'
 import Timestamp from './timestamp/timestamp'
-import Video from './video/video'
-import YoutubeVideo from './video/youtubeVideo'
 
 export {
   CodeSnippet,
